requirements.txt
numpy
Pillow
streamlit
